gcm clients now have separate instances.
Each instance contains internal http and xmpp client.
Xmpp client runs ping check and tries to reconnect after ping failure.
